refactor(dev): create remote real-time dev build and server
With this new way of development build and server, modify and save the extension code in the src to instantly see the reaction in the safari extension environment. The extension pages (popup, background etc.) are automatically refreshed, and the content scripts will instantly updated the next time the page is reload.